fix: move oauth config file
